Update `Fiber.defer`
the new method postpones releasing the connection to the next iteration of the event loop and only releases it if the file descriptor has changed since the last `io_wait` call. This ensures we are not releasing the connection until the DB query is completed (there can be several `io_wait` calls for one query)

The change is mostly caused by the bug in AR 7.0+ that makes the inital `Fiber.defer` approach unstable.

We also disable the method for AR 7.2+, as it doesn't hold onto connections anymore.