Correctly patch AR pool in Rails mode
we see that with some Rails applications `Rake` is defined even outside Rake tasks. Could be some dependencies always requiring it...

Anyway, we are safe to remove the `defined?` check, as we patch the AR pool only after the reactor is started, and it's not started in Rake tasks.

Additionally, this changes the way we notify the user of the patch to make sure we only notify when the reactor is started (i.e. when we actually patch the pool)