Ditch module builder pattern
the module builder pattern is an absolute mess from the implementation point of view. There're also several major issues with it, like `subscriber.is_a?(Rage::Events::Subscriber)` returning `false`.

The advantage of the module builder is just one explicit call - the module we are including is the subscription itself. On the other hand, IMO this is also a very unusual syntax for Ruby developers. And at this point I don't see much sense in using it.