Pass metadata as a `publish` argument
for the 27th time while working on this feature, I'm changing the approach. Instead of having `before_publish` callbacks, we will submit metadata as part of the `publish` call. We can always bring the callbacks back, but I think this solution is more flexible and makes more sense right now